requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
pillow
numpy
setuptools<82
einops
triton
pycocotools
psutil
transformers>=4.56.0
google-cloud-storage
opencv-python